fix(graph): note rules use the series' own pixel mapping — zoomed to a few milliseconds they drifted a millisecond's width right of their points
ChartPanel passed the view bounds to ChartNotes.byColumn cast to long; the
series use the exact doubles. In a 17 ms window that is up to 60 px, seen
on the owner's "Price by symbol" graph over MCP. byColumn(double, double,
int) now uses xToPx's formula; the long overload delegates.

ChartNotesTest.aRuleLandsOnItsPointWhenTheWindowHasFractionalBounds; the
truncating mutant is red there and on the colliding-columns test.
Round handoff for the reviewer: docs/handoff/handoff_analyser_finish_first_2026-09-16.md.

+++ src/main/java/telamin/fluxtion/audit/analyser/analyser/ui/ChartPanel.java
@@ -663,7 +663,7 @@ public final class ChartPanel extends JPanel {
         }
         Color rule = dark ? new Color(0x6E7681) : new Color(0x8C959F);
         Color pin = dark ? new Color(0xE3B341) : new Color(0x9A6700);
-        var columns = notes.byColumn((long) vx0, (long) vx1, plotW);
+        var columns = notes.byColumn(vx0, vx1, plotW);   // the exact view bounds, as the series use
         int index = 0;
         for (var entry : columns.entrySet()) {
             int px = plotX + entry.getKey();
